requirements.txt
pyyaml==6.0
acryl-datahub>=0.12.0
google-adk==0.0.1a15
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.2
python-dotenv==1.0.0
fastembed
qdrant-client 